requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database  
supabase==2.9.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Twilio
twilio==8.11.1

# WebSocket
websockets==12.0

# Cloud AI APIs (lightweight, fast, FREE tiers)
groq==0.36.0
openai==2.8.1

# Audio Processing (minimal)
pydub==0.25.1
numpy==1.26.3

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.20
piper-tts==1.3.0

# Logging
loguru==0.7.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
